+++ packages/jobs/src/__tests__/fixtures/src/jobs/cron.job.ts
@@ -1,20 +1,25 @@
+import * as z from 'zod'
+
 import { defineJob } from '../../../../index'
 
 export default defineJob({
+  schema: z.object({
+    message: z.string(),
+  }),
   cron: [
     {
       expression: '* * * * *',
       timezone: 'UTC',
-      payload: {
+      params: {
         message: 'from cron',
       },
     },
   ],
-  async handler({ payload, client, job, attempt }) {
+  async handler({ params, client, job, attempt }) {
     await client.raw(
       'INSERT INTO job_events (job_id, message, attempt) VALUES (?::uuid, ?, ?)',
       job.id,
-      (payload as { message: string }).message,
+      params.message,
       attempt,
     )
   },